[#87716776] Adds craft importer.py
The script first checks for the available parts in the current directory, reads the .craft
file, select the required parts (unique) and generates the corresponding blender objects as prefabs.
Then, each craft part is created by duplicating the corresponding unique prefab and setting the good transform.
To avoid overriding issues, material, textures and images are renamed accordingly to the prefab that uses them.