[#92418938] Avoids mu textures to be overridden by unpacking
In osgexport, packed textures are unpacked into files whose names are the basename of the
blender image's filepath. Since mu textures have the same name for each parts (model000.png etc),
we must rename the basename of the blender image's path to avoid the overridding at the unpacking step.
This issues concerns only PNG or TGA textures since the loading of a MBM generates a new blender image with
empty path.