[#92418938] Avoids loading twice the same texture
When loading a craft, parts are loaded sequencially. It happens that
two different parts use the same image, we check that the image has not been loaded previously.
If the Blender collection has an image whose filepath is the same as the current image (to be loaded),
we don't load the current image and don't create a texture for it.